moved some function to utils
